Some problems fixed
Simple structure is created
Unnecessary configs are removed
