+++ client/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Component from './Components/Component.js';
+
+ReactDOM.render(<Component />, document.getElementById('app'));
